Change Mapping

+++ Sources/Models/Occurrence.swift
@@ -2,12 +2,12 @@ import Foundation
 
 struct Occurrence: Codable {
     var id: String?
-    var date: Int64?
+    var date: Double?
     var title: String?
     var price: Double?
     var image: String?
-    var longitude: Int64?
-    var latitude: Int64?
+    var longitude: Double?
+    var latitude: Double?
     var description: String?
     var people: [People]?
     var cupons: [Cupon]?
